compile fix
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3034 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/IPC_HLE/WII_IPC_HLE_Device_sdio_slot0.cpp
@@ -295,7 +295,7 @@ u32 CWII_IPC_HLE_Device_sdio_slot0::ExecuteCommand(u32 _BufferIn, u32 _BufferInS
 
 			for (u32 i = 0; i < size; ++i)
 			{
-				(u8)buffer[i] = Memory::Read_U8(req.addr++);
+				buffer[i] = Memory::Read_U8(req.addr++);
 			}
 
 			size_t nWritten = fwrite(buffer, req.bsize, req.blocks, m_Card);
